Engine.Ioxide: move TLS transport plumbing into the engine
TlsDuplexPipe (the decrypt pump + kTLS-TX pipe adapter) and a new IoxideTls helper (StartService /
AcceptAsync) move into the engine so apps stop copy-pasting the transport glue; the app keeps only
cert/key/port config and the connectionFactory port decision. Engine gains an ioxide.tls dependency.
