+++ Engine/Ioxide/GenHTTP.Engine.Ioxide.csproj
@@ -8,9 +8,14 @@
     <ProjectReference Include="..\..\API\GenHTTP.Api.csproj" />
     <ProjectReference Include="..\Shared\GenHTTP.Engine.Shared.csproj" />
     <PackageReference Include="ioxide" Version="0.0.8" />
+    <PackageReference Include="ioxide.tls" Version="0.0.8" />
     <PackageReference Include="Glyph11" Version="0.3.5" />
     <!-- Parallel parse path for benchmarking — GENHTTP_IOXIDE_PARSER=pico (see ConnectionDriver). -->
     <PackageReference Include="Glyph11.Pico" Version="0.0.1" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="FIles\" />
   </ItemGroup>
 
 </Project>
